panier/produits

+++ src/main/java/com/shoploc/shoploc/domain/type/TypeProductRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TypeProductRepository extends JpaRepository<TypeProduct,Long> {
+public interface TypeProductRepository extends JpaRepository<TypeProduct,Integer> {
 
-    Optional<TypeProduct> findById (Long id);
+    Optional<TypeProduct> findById (Integer id);
 }
